requirements_test.txt
pytest==7.4.4
pytest-homeassistant-custom-component==0.13.109
pytest-asyncio==0.23.5
fpdf2==2.7.8
mypy==1.8.0
flake8==7.0.0
